BARK-87: Rename module dog-profile module to dog-profile-models
We will consolidate DogProfile and SubProfile into this models module.
