+++ src/app/user/(logged-in)/my-pets/_lib/components/sub-profile-form-controller.tsx
@@ -2,7 +2,7 @@
 
 import { BarkFormOption } from "@/components/bark/bark-form-option";
 import { SubProfileForm } from "./sub-profile-form";
-import { DogProfile } from "@/lib/bark/models/dog-profile";
+import { DogProfile } from "@/lib/bark/models/dog-profile-models";
 import { SubProfile } from "@/lib/bark/models/sub-profile";
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { useRouter } from "next/navigation";
