+++ src/app/user/(logged-in)/my-pets/_lib/components/sub-profile-form.tsx
@@ -10,7 +10,7 @@ import { BarkFormRadioGroup } from "@/components/bark/bark-form-radio-group";
 import { BarkFormOption } from "@/components/bark/bark-form-option";
 import { YesNoSchema } from "@/lib/bark/enums/yes-no";
 import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no-unknown";
-import { DogProfile } from "@/lib/bark/models/dog-profile";
+import { DogProfile } from "@/lib/bark/models/dog-profile-models";
 import { SubProfile, SubProfileSchema } from "@/lib/bark/models/sub-profile";
 import { Result } from "@/lib/utilities/result";
 import { zodResolver } from "@hookform/resolvers/zod";
