+++ src/lib/bark/models/encrypted-dog-profile.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { DogProfileSchema } from "./dog-profile";
+import { DogProfileSchema } from "./dog-profile-models";
 
 export const EncryptedDogProfileSchema = DogProfileSchema.omit({
   dogName: true,
